Add required_keys! helper to env
This allows you to do an aggressive post launch test to confirm all the
keys your apps need are present and ready to go.

It will raise an exception if any are missing and it will tell you
which of the required keys are missing.

Use this in rake tasks so cap can bail mid deploy, or in production.rb
to stop server boots.